Introduce basic error handling infrastructure
This change introduces the basic error handling infrastructure to the
program. For lack of stabilization of the process::Termination we work
fairly hard to make do with just returning a Result from main. To
accommodate that we introduce an ExitError struct whose main purpose is
to provide a proper error representation to the user.