Add support for retrieving a terminal device handle
The main logic of the program revolves around finding the proper Neovim
instance to work with. The way we identify this instance is by finding
all processes associated with a given (virtual) terminal. A path to a
terminal (e.g., /dev/pts/10) is what we assume to get from the outside.
With this change we validate this path to make sure that it indeed
represents a terminal and then return its device ID.