Introduce error context
System calls and most functions from the standard library return errors
that do not provide much value without additional context. A typical
error return such as "no such file or directory", for example, contains
neither the file that was not found nor the context in which the lookup
was performed, making it hard for the user to act on the error.
To counter this problem this change introduces the concept of an error
context. The existence of such a context can be enforced by virtue of a
function's signature and it is basically a string providing additional
information when emitting the error to the user.