Add logic to iterate over entries in /proc
The first goal the program needs to accomplish is to find the Unix
domain socket for the Neovim instance of interest. In order to do so, we
first need to deduce which process is the process "of interest".
The only way we have found to go about this task is by iterating over
the set of available processes and filtering out the relevant one(s). We
will do so in a piecemeal fashion: applying more and more filters until
we are certain that the remaining process is the one we were looking
for.
To that end, this change introduces the means to iterate over the
entries in the /proc virtual file system. Since different information is
contained in this directory and we are only interested in entries
representing processes, we apply a first filtering step to only yield
actual process IDs.