Filter out own process
The heuristic we will ultimately apply to find the Neovim instance of
interest is a prefix match against "nvim". Because our own program's
name begins with that same prefix and we are sure that it is not the
process we are looking for, this patch adds logic to filter out this
process from the list.