Filter processes by matching controlling terminals
The next step we use to determine the Neovim instance we are looking for
is to make filter out those processes that belong to the terminal the
user is looking in (i.e., has supplied to the program). The process'
terminal is the first file descriptor for the process' /proc entry,
e.g., /proc/<pid>/fd/0, and so we check whether that matches the
supplied one.