Map Neovim candidates to socket inodes
With the most recent filtering steps applied, we are in possession of a
list of Neovim instances. However, we are not so much interested in the
process IDs of those instances, but in the Unix domain sockets we can
use to connect to them.
As a first step to deducing the paths to those, this change iterates
over the process' file descriptors and filters out those that are of
type socket.