Retrieve UNIX domain sockets for found Neovim instances
The most recently added pipeline stage yielded the inodes of socket
files that a process has open. Not all of those sockets are necessarily
of interest to us. We are looking for a Unix domain socket, and in fact
the first of those, because the assumption is that while a Neovim
instance may have more than one, it will create the one that can be used
to control it first.
With this patch we map the inode we get passed in to Unix domain
sockets. The only viably way we found for performing this mapping is by
parsing /proc/net/unix and filtering and matching the entries in there
against the inode supplied.