Evaluate filtering results and indicate success to the user
Given the previously established filtering pipeline we are now able to
find the Neovim instances running in a certain terminal and map those
processes to Unix domain socket paths that we ultimately can connect to
for control purposes.
What we did not take care of so far is to evaluate the result of the
filtering. That is, while we did report any errors we encountered during
the filtering process, we could still end up with no instances or no
Unix domain sockets found.
With this change we add proper error reporting for those error cases.