Introduce Command enum and add 'find-socket' command
In the future we plan to have more than a single use case for the
program. That is, the user will be able to request different operations
to be performed.
To enable this use case, this change introduces a first recognized
command, find-socket, along with an enum that we use to represent the
individual supported commands in the program.