Add support for changing windows
This change introduces a second command, change-window. This command is
the first one to actually connect to and interact with the Neovim
instance found. We use the neovim-lib crate for sending data to the
process. The newly introduced command sends a sequence of keys to
Neovim, with the intention that those will change the currently active
window inside the editor.