Exit with an exit code indicating success of the window change
In order to build an end-to-end solution it is essential to understand
whether the window-change command actually resulted in a different
window being activated or so.
To that end, this change evaluates the currently active window before
and after sending the user-defined sequence of keys to the instance. If
the active window changed, we exit the program indicating success,
otherwise we indicate failure.