Ignore stopped nvim instances
With this change we check the status of any found nvim instance and
filter out those that are in stopped state.
The main intention of this change is to not attempt to send any data to
a stopped instance, as it cannot respond. So what happened so far was
that the send operation would ultimately time out, which is a bad user
experience. By filtering out stopped instances we circumvent this
problem.