Make find_self() return an OsString
The find_self() function currently returns a PathBuf. That's not
incorrect, considering that this is in fact the return value of the
internally used std::fs::read_link(). However, we know the /proc/self is
basically just a "relative" symbolic link to an entry representing a PID
in the same directory -- i.e., it is more akin to a file name. File
names are generally represented as OsString in Rust and so that's what
we adjust find_self() to return as well.