Adjust a few functions to accept a Path instead of an OsStr
A few of our functions are meant to be invoked on a path. However, the
argument type OsStr is typically used to represent a file name. Paths
have their own dedicated type in Rust.
With this change we switch them to using std::path::Path as the input
parameter's type.