Properly format the source code using rustfmt
In early revisions of the code we were not yet as strict about code
formatting as we should have been. As a result, a few parts of the
source code are not properly formatted as per rustfmt.
With this change we fix this flaw, ensuring that we can programmatically
enforce rustfmt compliance moving forward.