Adjust filter_nvim() to handle handle permission gracefully
With upcoming changes we will no longer only filter Neovim instances
after already checking for the expected TTY, but rather do so on
additional other processes.
As it turns out, though, we cannot read the exe symbolic link in a
process' proc entry if we don't own the process. Rather, this link
dereference will result in a permission error.
With this change we handle these errors gracefully, merely ignoring the
entry as it won't be the expected one anyway.