Adjust logic to work with Neovim 0.9.0
With version 0.9.0 (and commits 24488169564c ("feat(tui): run TUI as
external process") and 43e8ec92de9e ("fix(tui): more work in the TUI")
specifically) Neovim moved TUI handling into a dedicated process. That
breaks our TTY-to-socket-path-mapping logic, because this dedicated
process no longer has a TTY associated with it.
This change adjusts the socket discovery logic in such a way that we
first identify the Neovim process with the provided terminal and then
find another Neovim process that has this process as a parent. The
remainder of the existing logic remains the same.