Fix minimum supported Rust version CI job
The minimum supported Rust version CI job currently fails, because it
uses a new version of Cargo for generating the lock file and Rust 1.57,
which is our minimum supported Rust version, does not understand this
format.
Switch to using an older version of Rust for generating the lock file to
fix this issue.