Detect blocked Neovim instances
The Neovim RPC functions that we invoke can block basically indefinitely
if the editor is waiting for user input in a "modal" fashion. The
neovim-lib crate that we use by default uses a 5s timeout and then
errors out, so we will just eventually hit this time out, but the
behavior is super annoying.
With this change we invoke the nvim_get_mode "fast" function, which is
handled asynchronously by Neovim and, hence, is not running into any
timeout issues. If it reports that Neovim is currently blocked then we
bail out directly, reporting an error. Otherwise we continue as before.